+++ test/libsolidity/semanticTests/functionCall/named_args.sol
@@ -1,8 +1,10 @@
 contract test {
     function a(uint a, uint b, uint c) public returns (uint r) { r = a * 100 + b * 10 + c * 1; }
     function b() public returns (uint r) { r = a({a: 1, b: 2, c: 3}); }
+    function c() public returns (uint r) { r = a({b: 2, c: 3, a: 1}); }
 }
 // ====
 // compileViaYul: also
 // ----
 // b() -> 123
+// c() -> 123